requirements.txt
apscheduler
Django
rsa
djangorestframework
MySQL-python
django_apscheduler
